QA tools for comparison of packages in MetaCPAN and PAUSE
bin/qa/02packages outputs text in the style of PAUSE's
02packages.details.txt, but describes MetaCPAN's view of CPAN.  A key
difference is that versions are _always_ numified and formatted as
decimals with 12 digits of precision (4 repeats of 3 digit fixed-width
values).

bin/qa/normalize-02packages transforms PAUSE's 02packages in order to
ease direct textual diffing of it with the output of bin/qa/02packages
above.  Notably, this numifies version numbers in the same consistent
way, as PAUSE's indexing has changed over the years leading to a mix of
version formats in the file.

I'm using these tools to help identify classes of data mismatches
between MetaCPAN and PAUSE, which is useful for knowing what we have and
what we can rely on.